+++ packages/create-component/scaffolders/form-field/utils.js
@@ -1,11 +1,6 @@
 import fs from 'fs'
 import path from 'path'
-
-import ejs from 'ejs'
-
-import * as prettier from 'prettier'
-
-// import { prettierConf } from "./constants.js";
+import latestVersion from 'latest-version'
 
 const readFileContentRelativeToCurrentFile = (relativePath) => {
     const currentModuleFile = new URL(import.meta.url).pathname
@@ -21,38 +16,6 @@ const readFileContentRelativeToCurrentFile = (relativePath) => {
     }
 }
 
-function getAllFilePathsRecursively(directoryPath) {
-    const currentModuleFile = new URL(import.meta.url).pathname
-    const currentModuleDirectory = path.dirname(currentModuleFile)
-    const absolutePath = path.resolve(currentModuleDirectory, directoryPath)
-
-    const filesList = []
-
-    function readFilesHelper(currentPath, parentDirectory = '') {
-        const files = fs.readdirSync(currentPath)
-
-        files.forEach((file) => {
-            const filePath = path.join(currentPath, file)
-            const stats = fs.statSync(filePath)
-
-            if (stats.isDirectory()) {
-                // If it's a directory, recursively read its files
-                readFilesHelper(filePath, path.join(parentDirectory, file))
-            } else {
-                // If it's a file, add an object with fileName, filePath, and relativePath to the list
-                filesList.push({
-                    fileName: file,
-                    fileFullPath: filePath,
-                    relativeDirectoryPath: parentDirectory,
-                })
-            }
-        })
-    }
-
-    readFilesHelper(absolutePath)
-    return filesList
-}
-
 const getFileContentUsingFullPath = (filePath) => {
     try {
         // Read the file content synchronously
@@ -63,50 +26,20 @@ const getFileContentUsingFullPath = (filePath) => {
     }
 }
 
-const renderFileWithEjs = (fileContent, data) => {
-    return ejs.render(fileContent, data)
-}
-
-function writeFileWithFolderCreation(targetFilePath, renderedFileContent) {
-    const targetDirectory = path.dirname(targetFilePath)
-
-    // Create the directory if it doesn't exist
-    if (!fs.existsSync(targetDirectory)) {
-        fs.mkdirSync(targetDirectory, { recursive: true })
-    }
-
-    // Write the file content to the target file
-    fs.writeFileSync(targetFilePath, renderedFileContent, 'utf-8')
-}
-
-const formatWithPrettier = async (content, filePath) => {
+const getLatestPackageVersion = async (packageName) => {
     try {
-        const formattedContent = await prettier.format(content, {
-            filepath: filePath,
-        })
-        return formattedContent
-    } catch (err) {
-        return content
+        const verison = await latestVersion(packageName)
+        return verison
+    } catch (error) {
+        console.log(
+            chalk.red(`Error retrieving the latest version: ${error.message}.`)
+        )
+        process.exit(0)
     }
-}
-
-function isValidPackageName(projectName) {
-    return /^(?:@[a-z\d\-*~][a-z\d\-*._~]*\/)?[a-z\d\-~][a-z\d\-._~]*$/.test(
-        projectName
-    )
-}
-
-const makeDirectory = (projectPath) => {
-    fs.mkdirSync(projectPath)
 }
 
 export {
-    getAllFilePathsRecursively,
     getFileContentUsingFullPath,
-    renderFileWithEjs,
-    writeFileWithFolderCreation,
-    formatWithPrettier,
     readFileContentRelativeToCurrentFile,
-    isValidPackageName,
-    makeDirectory,
+    getLatestPackageVersion,
 }
